+++ dashboard/src/actions/types.js
@@ -9,3 +9,6 @@ export const NETWORK_ERROR = "NETWORK_ERROR";
 
 /*USER AUTHENTICATION */
 export const KEEP_USER_LOGGED_IN = "KEEP_USER_LOGGED_IN";
+export const USER_NOTION_ALERTS = "USER_NOTION_ALERTS";
+export const SET_LOGIN_BUTTON = "SET_LOGIN_BUTTON";
+export const SET_SIGNUP_BUTTON = "SET_SIGNUP_BUTTON";
