+++ dashboard/src/modules/components/AuthComponent/index.less
@@ -22,6 +22,9 @@
         .pf-c-card {
             padding: 2vh 0;
         }
+        .continueBtn {
+            padding: inherit;
+        }
         .signup-card {
             height: 80%;
             overflow: scroll;
