+++ dashboard/src/reducers/authReducer.jsx
@@ -1,7 +1,13 @@
-import { KEEP_USER_LOGGED_IN } from "../actions/types";
+import { KEEP_USER_LOGGED_IN,
+    USER_NOTION_ALERTS,
+    SET_SIGNUP_BUTTON,
+    SET_LOGIN_BUTTON} from "../actions/types";
 
 const initialState = {
-    keepLoggedIn : false
+    keepLoggedIn : false,
+    alerts: [],
+    isLoginBtnDisabled: true,
+    isSignupBtnDisabled: true,
 }
 
 const AuthReducer = (state = initialState, action = {}) => {
@@ -11,6 +17,21 @@ const AuthReducer = (state = initialState, action = {}) => {
             return {
                 ...state,
                 keepLoggedIn: payload
+            }
+        case USER_NOTION_ALERTS:
+            return {
+                ...state,
+                alerts: [...payload]
+            }
+        case SET_LOGIN_BUTTON:
+            return {
+                ...state,
+                isLoginBtnDisabled: payload
+            }
+        case SET_SIGNUP_BUTTON:
+            return {
+                ...state,
+                isSignupBtnDisabled: payload
             }
         default:
             return state;
