+++ dashboard/src/utils/apiConstants.js
@@ -1,2 +1,4 @@
 export const LOCAL_BASE_URL = "http://10.1.170.201/";
 
+export const REGISTER = "api/v1/register";
+export const LOGIN = "api/v1/login"
